+++ changelog.md
@@ -7,6 +7,7 @@ This version is not released yet and is under active development.
 ```
 
 - \[paru\] Add `paru` support.
+- \[mpm\] Colorize version differences in `outdated` operation output.
 - \[mpm\] Add manager homepage URL metadata.
 
 ## {gh}`5.2.0 (2022-06-16) <compare/v5.1.0...v5.2.0>`